+++ R/estimatepopbm.R
@@ -244,12 +244,15 @@ clusterize_unipartite_networks <- function(netlist,
 #' @param nb_run An integer, the number of run the algorithm do.
 #' @param global_opts Global options for the outer algorithm and the output
 #' @param fit_opts Fit options for the VEM algorithm
-#' @param fit_init Do not use!
-#' Optional fit init from where initializing the algorithm.
+#' @param partition_init Optional partition list, a list of fitted collections
+#' (bisbmpop) from which to start partitioning
+#' @param full_collection_init Optional full collection, a bisbmpop object
+#' containing the fit of all the networks
 #' @param full_inference The default "FALSE", the algorithm stop once splitting
 #' groups of networks does not improve the BICL criterion. If "TRUE", then
 #' continue to split groups until a trivial classification of one network per
 #' group.
+#' @inheritParams stats::hclust
 #' @param verbose A boolean, should the function be verbose or not. Default to
 #' TRUE.
 #'
@@ -299,8 +302,10 @@ clusterize_bipartite_networks <- function(netlist,
                                           nb_run = 3L,
                                           global_opts = list(),
                                           fit_opts = list(),
-                                          fit_init = NULL,
+                                          partition_init = NULL,
+                                          full_collection_init = NULL,
                                           full_inference = FALSE,
+                                          method = "single",
                                           verbose = TRUE,
                                           temp_save_path = tempfile(fileext = ".Rds")) {
   check_bipartite_colsbm_models(colsbm_model = colsbm_model)
@@ -320,28 +325,50 @@ clusterize_bipartite_networks <- function(netlist,
   fit_opts <- fo
 
   # Fit the initial model on the full collection
-  if (verbose) {
-    if (!is.null(temp_save_path)) {
-      cli::cli_alert_info("A save file will be created at {.val {temp_save_path}} and updated after each step")
-    }
-    cli::cli_h1("Fitting the full collection")
+  if (verbose & !is.null(temp_save_path)) {
+    cli::cli_alert_info("A save file will be created at {.val {temp_save_path}} and updated after each step")
   }
   start_time <- Sys.time()
-  my_bisbmpop <- estimate_colBiSBM(
-    netlist = netlist,
-    colsbm_model = colsbm_model,
-    net_id = net_id,
-    distribution = distribution,
-    nb_run = nb_run,
-    global_opts = global_opts,
-    fit_opts = fit_opts
-  )
+  if (is.null(full_collection_init)) {
+    cli::cli_h1("Fitting the full collection")
 
-  clustering_queue <- list(my_bisbmpop)
-  list_model_binary <- list()
-  cluster <- rep(1, length(netlist))
-  names(cluster) <- net_id
-  clustering_history <- as.data.frame(matrix(cluster, nrow = 1L))
+    my_bisbmpop <- estimate_colBiSBM(
+      netlist = netlist,
+      colsbm_model = colsbm_model,
+      net_id = net_id,
+      distribution = distribution,
+      nb_run = nb_run,
+      global_opts = global_opts,
+      fit_opts = fit_opts
+    )
+  } else {
+    # If a full collection is provided, use it to initialize the clustering
+    cli::cli_h1("Using a provided full collection to initialize the clustering")
+    stopifnot("full_collection_init should be a bisbmpop object" = is(full_collection_init, "bisbmpop"), "full_collection$A should be identical to netlist" = identical(full_collection_init$A, netlist))
+    my_bisbmpop <- full_collection_init
+  }
+
+  if (is.null(partition_init) || is(partition_init, "bisbmpop")) {
+    clustering_queue <- list(my_bisbmpop)
+    list_model_binary <- list()
+    cluster <- rep(1, length(netlist))
+    names(cluster) <- net_id
+    clustering_history <- as.data.frame(matrix(cluster, nrow = 1L))
+  } else {
+    if (verbose) {
+      cli::cli_inform("A list of fits was provided, the clustering will start from this list")
+    }
+    # Starting from a list of fits
+    stopifnot("fit_init should be a list of bisbmpop objects" = is(partition_init, "list"))
+    clustering_queue <- partition_init
+    list_model_binary <- list()
+    cluster <- unlist(lapply(seq_along(clustering_queue), function(i) {
+      rep(i, length(clustering_queue[[i]]$net_id))
+    }))
+    stopifnot("Cluster vector should be the same length as net_id" = length(cluster) == length(net_id))
+    names(cluster) <- net_id
+    clustering_history <- as.data.frame(matrix(cluster, nrow = 1L))
+  }
 
   if (verbose) {
     cli::cli_h1("Beginning clustering")
@@ -370,6 +397,7 @@ clusterize_bipartite_networks <- function(netlist,
     cl <- partition_networks_list_from_dissimilarity(
       networks_list = fit$A,
       dissimilarity_matrix = dist_bm,
+      method = method,
       nb_groups = 2L
     )
 
@@ -378,7 +406,7 @@ clusterize_bipartite_networks <- function(netlist,
     }
     # Fit models for the sub-collections
     fits <- future.apply::future_lapply(
-      c(1, 2),
+      seq(1, length(unique(cl))),
       function(k) {
         Z_init <- lapply(
           seq_along(fit$model_list),
@@ -454,8 +482,12 @@ clusterize_bipartite_networks <- function(netlist,
     partition = list_model_binary,
     cluster = cluster,
     elapsed_time = Sys.time() - start_time,
-    clustering_history = clustering_history
+    clustering_history = clustering_history,
+    full_collection = my_bisbmpop
   )
+  if (verbose) {
+    cli::cli_alert_info("After clustering the partition has a BIC-L of {.val {compute_bicl_partition(output_list$partition)}}")
+  }
   if (!is.null(temp_save_path)) {
     saveRDS(output_list, temp_save_path)
     if (verbose) {
@@ -590,6 +622,7 @@ compute_dissimilarity_matrix.bmpop <- function(
 #' @param networks_list The list of networks to partition
 #' @param dissimilarity_matrix The dissimilarity matrix computed with
 #' `compute_dissimilarity_matrix`.
+#' @inheritParams stats::hclust
 #' @param nb_groups An integer, the number of groups. Defaults to 2
 #'
 #' @return A vector, eventually named according to `networks_list` names.
@@ -598,6 +631,7 @@ compute_dissimilarity_matrix.bmpop <- function(
 partition_networks_list_from_dissimilarity <- function(
     networks_list,
     dissimilarity_matrix,
+    method = "single",
     nb_groups = 2L) {
   # Sanity checks
   check_networks_list(networks_list)
@@ -616,12 +650,7 @@ partition_networks_list_from_dissimilarity <- function(
   if (M >= 3L) {
     # If there is more than 3 networks they are splitted using partition around
     # K-medioids
-    cl <- cluster::pam(
-      x = sqrt(dissimilarity_matrix),
-      k = nb_groups,
-      diss = TRUE,
-      cluster.only = TRUE
-    )
+    cl <- cutree(hclust(as.dist(sqrt(dissimilarity_matrix)), method = method), k = nb_groups)
   } else {
     cl <- c(1L, 2L)
   }
@@ -731,10 +760,6 @@ clusterize_bipartite_networks_graphon <- function(
   if (!is.null(fit_init) & verbose) {
     cli::cli_alert_info("Starting from a list of fits")
     cli::cli_alert_warning("This feature is still experimental and cluster vector may not be accurate\n\n")
-  }
-
-  compute_bicl_partition <- function(collections) {
-    return(sum(sapply(collections, function(col) col$best_fit$BICL)))
   }
 
   # Historique des fusions